moved PaperFormat to own module
